requirements.txt
torch
torchvision
diffusers
transformers
accelerate
datasets
xarray
h5netcdf
hydra-core
wandb
wandb[media]